aulas/gobarber/backend/src/modules/users/services/UpdateUserAvatarService.spec.ts
import AppError from '@shared/errors/AppError';

import FakeUsersRepository from '@modules/users/repositories/fakes/FakeUsersRepository';
import FakeStorageProvider from '@shared/container/providers/StorageProvider/fakes/FakeStorageProvider';

import UpdateUserAvatarService from './UpdateUserAvatarService';

let fakeUsersRepository: FakeUsersRepository;
let fakeStorageProvider: FakeStorageProvider;
let updateUserAvatarService: UpdateUserAvatarService;

describe('UpdateUserAvatar', () => {
  beforeEach(() => {
    fakeUsersRepository = new FakeUsersRepository();
    fakeStorageProvider = new FakeStorageProvider();
    updateUserAvatarService = new UpdateUserAvatarService(
      fakeUsersRepository,
      fakeStorageProvider,
    );
  });

  it('atualizar avatar de um usuário existente', async () => {
    const user = await fakeUsersRepository.create({
      name: 'John Doe',
      email: 'email@email.com',
      password: '123456',
    });

    const avatarUser = await updateUserAvatarService.execute({
      user_id: user.id,
      avatarFilename: 'imagem.jpg',
    });

    expect(avatarUser).toHaveProperty('id');
  });

  it('não atualizar avatar de um usuário inexistente', async () => {
    expect(
      updateUserAvatarService.execute({
        user_id: 'non-existing-user',
        avatarFilename: 'imagem.jpg',
      }),
    ).rejects.toBeInstanceOf(AppError);
  });

  it('deletar um avatar antigo ao incluir um novo', async () => {
    const deleteFile = jest.spyOn(fakeStorageProvider, 'deleteFile');

    const user = await fakeUsersRepository.create({
      name: 'John Doe',
      email: 'email@email.com',
      password: '123456',
    });

    await updateUserAvatarService.execute({
      user_id: user.id,
      avatarFilename: 'imagem1.jpg',
    });

    await updateUserAvatarService.execute({
      user_id: user.id,
      avatarFilename: 'imagem2.jpg',
    });

    expect(deleteFile).toHaveBeenCalledWith({ file: 'imagem1.jpg' });
    expect(user.avatar).toBe('imagem2.jpg');
  });
});
